+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixes
 * Added missing function to retrieve inscopix metadata [PR #436](https://github.com/catalystneuro/roiextractors/pull/436)
 ### Deprecations And Removals
+* The `get_image_size()` method in SegmentationExtractor is deprecated and will be removed on or after January 2025. Use `get_frame_shape()` instead. [PR #TBD](https://github.com/catalystneuro/roiextractors/pull/TBD)
+* The `get_num_frames()` method in SegmentationExtractor is deprecated and will be removed on or after January 2025. Use `get_num_samples()` instead. [PR #TBD](https://github.com/catalystneuro/roiextractors/pull/TBD)
 
 ### Improvements
 * Bruker series can now read sequences of type `BrightnessOverTime` [PR #448](https://github.com/catalystneuro/roiextractors/pull/448)